Create initial tests
This test rig creates a mock "conntrack" executable and puts it in $PATH then
checks if the exporter generates metrics accordingly from it.

Any changes in the code /should/ pass the tests.  If the tests fail after
labels were added, then that's a bug of the tests.  By accident the count
metric has a `cpu` label; removing that should be fine, but may still need
fixes of the test rig.