Change the API surface of exporter package
While it is good practice to use a subpackage from the start and use its API
surface in a minimal main package that just does configuration and
starting/stopping of resources, I was not smart in chosing how that API surface
should look like.

This kind-of breaking change removes the coupling to the ,,prometheus client
library'' and reduce the package API to what IMHO should matter: There's an
HTTP handler that somehow generates metrics in a standardized format.

While at it this introduces the functional options pattern.