Hack support for new field "ignore"
According to issue #19 there is a new field "ignore" in the output of the
conntrack tool.  My assumption is that some Linux kernels export this field and
the conntrack displays its data when available.  This commit hacks the regexp
to include that field to the metrics corpus.  Also add a regression test.

TODO: Make the exporter resilient against future unknown fields.